🐛 Tillat desimaltall som bompenger og fergjekostnader.

+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/dagligReise/beregning/privatBil/ReiseMedPrivatBil.kt
@@ -34,8 +34,8 @@ data class ReiseMedPrivatBilDelperiode(
     override val fom: LocalDate,
     override val tom: LocalDate,
     val reisedagerPerUke: Int,
-    val bompengerPerDag: Int?,
-    val fergekostnadPerDag: Int?,
+    val bompengerPerDag: Double?,
+    val fergekostnadPerDag: Double?,
 ) : Periode<LocalDate>,
     KopierPeriode<ReiseMedPrivatBilDelperiode> {
     override fun medPeriode(
